jenkinsfile deployment if else

+++ Jenkinsfile
@@ -16,20 +16,21 @@ pipeline {
       steps {
       sh "docker build -t caprearch/helloworld-war:${env.BUILD_ID} ."
       sh "docker tag caprearch/helloworld-war:${env.BUILD_ID} caprearch/helloworld-war:latest"
-	  sh "docker service ls"
-	  sh "docker service rm helloworld-war"
-	  sh "docker service create --name helloworld-war --replicas 1 --publish 9797:9797 caprearch/helloworld-war:${env.BUILD_ID}"
-	  sh "docker service ls"
 	  
       }
     }
 	
 	stage('Docker Deployment') {
       steps {
-	  sh "docker service ls"
-	  sh "docker service rm helloworld-war"
-	  sh "docker service create --name helloworld-war --replicas 1 --publish 9797:9797 caprearch/helloworld-war:${env.BUILD_ID}"
-	  sh "docker service ls"
+	   script {
+			sh "docker service ls"
+				if (helloworld-war) {
+				sh "docker service rm helloworld-war"
+				} else {
+				sh "docker service create --name helloworld-war --replicas 1 --publish 9797:9797 caprearch/helloworld-war:${env.BUILD_ID}"
+				sh "docker service ls"
+				}
+			}
 	  
       }
     }
